Add error checking and fix intended movetodir behavior

+++ shell.c
@@ -248,13 +248,31 @@ void executeCommand(Command *pcommand){
 // raises an error
 void movetodir(Command *pcommand){
 
+    // Do nothing when told to do nothing
+    if(pcommand->numOfParameters == 0) return;
+
     char* newDirectory = pcommand->parameters[0];
+    if(strcmp(newDirectory, "..") == 0){
+        // If we enter the directory as '..' we will go back one directory
+        // Effectivly we find the pointer to the last slash in the directory string
+        // and set it to the termination character.
+        char *last_slash = strrchr(dirinfo->currentDirectory, '/');
+
+        // Couldn't go back
+        if(last_slash == NULL) {
+            printf("Could not desend directory tree.\n");
+        }
+
+        *last_slash ='\0';
+        return;
+    }
     DIR* dir = opendir(newDirectory);
 
     if(dir != NULL){
         closedir(dir);
-        dirinfo->currentDirectory = newDirectory;
-
+        //dirinfo->currentDirectory = newDirectory;
+        strcat(dirinfo->currentDirectory, "/");
+        strcat(dirinfo->currentDirectory, newDirectory);
         if(DEBUGMODE){
             printf("Changed to: %s\n", dirinfo->currentDirectory);
         }
@@ -322,7 +340,10 @@ void byebye(){
 // Replays the given command referenced in the order that they are printed (0 
 // is the most recent call)
 void replay(Command *pcommand){
-    
+
+    // Do nothing when told to do nothing
+    if(pcommand->numOfParameters == 0) return;
+
     char* strNumber = pcommand->parameters[0];
     int commandNumber = (int) strtol(strNumber, (char **)NULL, 10);
     
@@ -341,6 +362,10 @@ void replay(Command *pcommand){
 // Runs the given program with the parameters passed to it
 void start(Command *pcommand){
     char program[BUFFERSIZE];
+
+    // Do nothing when told to do nothing
+    if(pcommand->numOfParameters == 0) return;
+
     strcpy(program, pcommand->parameters[0]);
 
     //pids
@@ -368,6 +393,10 @@ void start(Command *pcommand){
 // Prints the PID of the processes that was created
 void background(Command *pcommand){
     char program[BUFFERSIZE];
+
+    // Do nothing when told to do nothing
+    if(pcommand->numOfParameters == 0) return;
+
     strcpy(program, pcommand->parameters[0]);
     char *cmdString[] = {program, NULL};
     pid_t parent = getpid();
@@ -386,6 +415,9 @@ void background(Command *pcommand){
 
 // Terminates the process with the given PID
 void dalek(Command *pcommand){
+    // Do nothing when told to do nothing
+    if(pcommand->numOfParameters == 0) return;
+    
     int pid = strtol(pcommand->parameters[0], (char **)NULL, 10);
     printf("Killing [%d]\n", pid);
     kill(pid, SIGKILL);
